requirements.txt
flask
mindsdb
python-dotenv